Move some file for linter
